Clean up scripts for old website (#391)
