+++ scripts/generate_learning_path_markdown.js
@@ -1,6 +1,8 @@
 const { join, basename, relative } = require('path')
 const fs = require('fs')
 const yamlFront = require('yaml-front-matter')
+const asciidoctor = require('asciidoctor')()
+const YAML = require('yaml')
 
 const getContributors = require('./get_contributors')
 const mkdirSync = require('./mkdir_sync')
@@ -18,7 +20,17 @@ const languageMap = {
   ru: 'Russian'
 }
 
-module.exports = async (writeDir, generatorFn, workbookFn, isJekyllSite) => {
+const urls = YAML.parse(fs.readFileSync(join('..', 'config', 'urls.yaml'), 'utf-8'))
+
+const getYouTubeCode = (section, articleNumber) => {
+  const firstEntryOfGroupIndex = urls.findIndex(entry => entry.section === section.toLowerCase())
+  const currentPageIndexOffset = articleNumber - 1
+  const youtubeUrl = urls[firstEntryOfGroupIndex + currentPageIndexOffset].video.youtube
+  return youtubeUrl.replace('https://www.youtube.com/watch?v=', '')
+}
+
+(async () => {
+  const writeDir = 'newsite'
   mkdirSync(join('.', writeDir))
 
   sections.forEach((section, sectionDataIndex) => {
@@ -29,51 +41,50 @@ module.exports = async (writeDir, generatorFn, workbookFn, isJekyllSite) => {
 
     translations.concat('' /* The English original */).forEach(async (translation) => {
       const isTranslation = translation !== ''
-      const writePath = join(baseWritePath, translation)
       const readPath = join(baseReadPath, translation)
       const articles = getArticleFiles(readPath)
 
-      // Bad separation of concerns (Hugo/Jekyll specific logic shouldn't be here) but is temporary until Jekyll site is decomissioned
-      if (isJekyllSite) {
-        mkdirSync(writePath)
-      } else {
-        let indexReadPath = join(readPath, 'index.md')
-        const indexTranslated = fs.existsSync(indexReadPath)
-        if (!indexTranslated) {
-          // Default to English index page content
-          indexReadPath = join(baseReadPath, 'index.md')
-        }
-        const indexWritePath = isTranslation ? join(baseWritePath, `_index.${translation}.md`) : join(baseWritePath, '_index.md')
-
-        let { indexContent, ...indexFrontMatter } = yamlFront.loadFront(fs.readFileSync(indexReadPath, 'utf-8'), { contentKeyName: 'indexContent' })
-        indexFrontMatter.image = section.image
-        indexFrontMatter.contributors = await getContributors(relative('..', readPath))
-        indexFrontMatter.weight = sectionDataIndex
-
-        if (!indexTranslated) {
-          // Add placeholder text if index is not translated
-          indexContent = `This text will be translated soon to ${languageMap[translation]}.${indexContent}`
-        }
-
-        writeMarkdownFile(indexWritePath, indexFrontMatter, indexContent)
+      let indexReadPath = join(readPath, 'index.md')
+      const indexTranslated = fs.existsSync(indexReadPath)
+      if (!indexTranslated) {
+        // Default to English index page content
+        indexReadPath = join(baseReadPath, 'index.md')
       }
+      const indexWritePath = isTranslation ? join(baseWritePath, `_index.${translation}.md`) : join(baseWritePath, '_index.md')
+
+      let { indexContent, ...indexFrontMatter } = yamlFront.loadFront(fs.readFileSync(indexReadPath, 'utf-8'), { contentKeyName: 'indexContent' })
+      indexFrontMatter.image = section.image
+      indexFrontMatter.contributors = await getContributors(relative('..', readPath))
+      indexFrontMatter.weight = sectionDataIndex
+
+      if (!indexTranslated) {
+        // Add placeholder text if index is not translated
+        indexContent = `This text will be translated soon to ${languageMap[translation]}.${indexContent}`
+      }
+
+      writeMarkdownFile(indexWritePath, indexFrontMatter, indexContent)
 
       articles.forEach(async (article) => {
         const articleTitle = article.asciiDoc.match(/== (.*)/)[1]
         const articleNumber = basename(article.filePath).split('-')[0]
         const contributors = await getContributors(relative('..', article.filePath))
+        const fileName = isTranslation ? join(baseWritePath, [articleNumber, translation, 'md'].join('.')) : join(baseWritePath, [articleNumber, 'md'].join('.'))
+        const weight = parseInt(articleNumber)
+        const youtubeCode = getYouTubeCode(section.learning_path_group, weight)
 
-        generatorFn({
-          section,
-          articleTitle,
-          articleNumber,
-          isTranslation,
-          article,
-          translation,
+        const frontMatter = {
+          title: articleTitle,
           contributors,
-          writePath,
-          baseWritePath
-        })
+          image: `https://img.youtube.com/vi/${youtubeCode}/mqdefault.jpg`,
+          featured: weight === 1,
+          weight,
+          youtubeCode
+        }
+
+        const titleStripped = article.asciiDoc.replace(/== (.*)/, '')
+        const body = section.renderArticles || isTranslation ? asciidoctor.convert(titleStripped) : ''
+
+        writeMarkdownFile(fileName, frontMatter, body)
       })
 
       // Workbooks not translated.
@@ -82,13 +93,18 @@ module.exports = async (writeDir, generatorFn, workbookFn, isJekyllSite) => {
         const contributors = await getContributors(`workbook/${section.workbook}`)
         const workbookPosition = articles.length + 1
 
-        workbookFn({
-          section,
-          workbookFileName,
+        const workbookFrontMatter = {
+          title: 'Workbook',
           contributors,
-          workbookPosition
-        })
+          image: section.image,
+          weight: workbookPosition
+        }
+
+        const workbookReadPath = join('..', 'workbook', section.workbook)
+        const body = asciidoctor.convert(fs.readFileSync(workbookReadPath, 'utf-8'))
+
+        writeMarkdownFile(workbookFileName, workbookFrontMatter, body)
       }
     })
   })
-}
+})()
